Refactor : AWS, Base, Event  리팩토링
