+++ src/main/java/com/mypill/global/aws/s3/dto/AmazonS3Dto.java
@@ -13,5 +13,4 @@ public class AmazonS3Dto {
 
     private String cdnUrl;
     private String originUrl;
-
 }
